Forward rendering : First attempt at model loading (separate attributes)
On this first attemps, I tried to take advantage of the fact that tiny
obj extracts the vertices, normals and texcoords in separate arrays to
directly build three separate VBOs and bind them accordingly.

However, I could not get the binding to work as intended, and will fall
back on a more conventionnal solution : build an interleaved buffer from
the extracted data.