Forward rendering : Model Loading
Implemeted model loading with a data structure akin to simple geometry
for the purpose of building the VBO and IBOs.
The VBO is now interleaved, and built with the Vertex3f3f2f structure.

How much the VBO size and indexation could be optimized remains to be
seen, as I am for now using a direct and straighforward approach :
- build one vertex buffer for the whole model
- build one index buffer for each shape contained in the model
- currently, there is nothing done to prevent redundant vertex data to
be stored in the buffer