Defered Render : Implemented controls to displays GBuffers.
The background still has its odd coloring.
Also tried to display Depth Buffer in Shading Pass, but the result is
odd, it is probably linked to the Depth Buffer format.

Another idea could be to use one more texture in color mode only for
display, and write the gl_FragCoord.z from the Geometry Pass in it.
This could be also used as a back-up texture for the upcoming Shadow
Mapping, if the DepthBuffer really has a bigger problem thans display.