Defered Render : Modified the shader to remove strange colored backgroung using depth == 1.f
This solutions is merely a hack, and the light calculation should be
revewed to make the the problem disapear.

The upside is now I am 99% sure this small bug comes from the lighting
computation in my shader, which was adequate on forward rendering buit
might need some tweaking; and can make a confirmation the depth buffer
is working as expected, and that the bug isn t comming from somewhere
upstream in my code.

The definitive fix for this bug is now a low priority task.